requirements.txt
aiohttp>=3.9.5
aiosignal>=1.3.1
annotated-types>=0.6.0
anyio>=4.3.0
attrs>=23.2.0
certifi>=2024.2.2
charset-normalizer>=3.3.2
click>=8.1.7
fastapi>=0.110.1
frozenlist>=1.4.1
h11>=0.14.0
idna>=3.7
iniconfig>=2.0.0
Jinja2>=3.1.3
MarkupSafe>=2.1.5
multidict>=6.0.5
packaging>=24.0
pluggy>=1.4.0
pydantic>=2.7.0
pydantic_core>=2.18.1
pytest>=8.1.1
pytest-asyncio>=0.23.6
pytest-html>=4.1.1
pytest-metadata>=3.1.1
pytest-sugar>=1.0.0
python-dotenv>=1.0.1
requests>=2.31.0
sniffio>=1.3.1
starlette>=0.37.2
termcolor>=2.4.0
typing_extensions>=4.11.0
urllib3>=2.2.1
uvicorn>=0.29.0
websockets>=12.0
yarl>=1.9.4
